+++ src/plugins.ts
@@ -21,6 +21,7 @@ import eslintPluginPreferArrowFunctions from 'eslint-plugin-prefer-arrow-functio
 import eslintPluginPromise from 'eslint-plugin-promise';
 // TODO it returns undefined without `* as` syntax for some reason
 import * as eslintPluginQwik from 'eslint-plugin-qwik';
+import eslintPluginReact from 'eslint-plugin-react';
 import * as eslintPluginRegexp from 'eslint-plugin-regexp';
 import eslintPluginSecurity from 'eslint-plugin-security';
 import eslintPluginSonar from 'eslint-plugin-sonarjs';
@@ -79,4 +80,5 @@ export const ALL_ESLINT_PLUGINS: FlatConfigEntry['plugins'] & {} = {
   // @ts-expect-error types mismatch
   'json-schema-validator': eslintPluginJsonSchemaValidator,
   css: eslintPluginCss,
+  react: eslintPluginReact,
 };